Collapse BindListener. Only used in one place.
